First implementation of a custom cluster detector
This ROS2 package consists of a single subscriber-publisher node. Its aim is to find obstacle circles around the robot using lidar scanning.

The lidar scan is processed in the scan_callback function, in three parts:

-First, a median filter is applied to the scan to correct for noise.

-Secondly, points close to each other are grouped into clusters

-Thirdly, the center of the circle is assumed to be at the same distance as the furthest point in the cluster. This approximation is based on the theoretical assumption that we see a complete semicircle.
We can therefore deduce that the radius of the circle can be calculated by looking at the angle of vision it covers.

Finally, we send this data to a topic.

Signed-off-by: Thomas Bonnefille <thomas.bonnefille@outlook.com>