Patch back browsing of the scan
In the previous implementation, the code always browse the code backward
to joint the beginning and the end of the scan even if it is not always needed.

This patch avoid this problem by declaring that a point is in a cluster
if it is at a distance of less than 10m (instead of 10km).

Signed-off-by: Thomas Bonnefille <thomas.bonnefille@outlook.com>